# ProductAPI/src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "product-api"
servers:
  - url: "http://localhost:8090"
...
paths:
...
  /api/v1/perfumes/ids:
    post:
      tags:
        - "PerfumeController"
      summary: "Get perfumes by ids"
      operationId: "getPerfumesByIds"
      requestBody:
        content:
          'application/json':
            schema:
              type: "array"
              items:
                type: "integer"
                format: int64
              example: [1, 2, 3]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by filter parameters"
      operationId: "findPerfumesByFilterParams"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumers:
                  type: "array"
                  items:
                    type: "string"
                    example: "Dior"
                genders:
                  type: "array"
                  items:
                    type: "string"
                    example: "male"
                prices:
                  type: "array"
                  items:
                    type: "integer"
                    format: int32
                  example: [1, 200]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search/gender:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by gender"
      operationId: "findByPerfumeGender"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                genders:
                  type: "array"
                  items:
                    type: "string"
                    example: "male"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/search/perfumer:
    post:
      tags:
        - "PerfumeController"
      summary: "Find perfumes by brand"
      operationId: "findByPerfumer"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumers:
                  type: "array"
                  items:
                    type: "string"
                  example: ["Dior", "Creed"]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/perfumes/graphql/ids:
    post:
      tags:
        - "PerfumeController"
      summary: "Get list of perfumes by ids GraphQL query"
      operationId: "getPerfumesByIdsQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfumesIds(ids: [16,17,18]) {id perfumeTitle perfumer price filename} }"

      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    perfumeTitle:
                      type: "string"
                    perfumer:
                      type: "string"
                    price:
                      type: "integer"
                      format: int32
                    filename:
                      type: "string"
                  example:
                    id: 43
                    perfumeTitle: "Sauvage"
                    perfumer: "Dior"
                    price: 62
                    filename: "empty.jpg"
  /api/v1/perfumes/graphql/perfumes:
    post:
      tags:
        - "PerfumeController"
      summary: "Get list of perfumes by GraphQL query"
      operationId: "getAllPerfumesByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfumes {id perfumeTitle perfumer price filename} }"

      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: "integer"
                      format: int64
                    perfumeTitle:
                      type: "string"
                    perfumer:
                      type: "string"
                    price:
                      type: "integer"
                      format: int32
                    filename:
                      type: "string"
                  example:
                    id: 43
                    perfumeTitle: "Sauvage"
                    perfumer: "Dior"
                    price: 62
                    filename: "empty.jpg"
  /api/v1/perfumes/graphql/perfume:
    post:
      tags:
        - "PerfumeController"
      summary: "Get perfume by GraphQL query"
      operationId: "getPerfumeByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ perfume(id: 43) {id perfumeTitle perfumer price filename} }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  perfumeTitle:
                    type: "string"
                  perfumer:
                    type: "string"
                  price:
                    type: "integer"
                    format: int32
                  filename:
                    type: "string"
                example:
                  id: 43
                  perfumeTitle: "Sauvage"
                  perfumer: "Dior"
                  price: 62
                  filename: "empty.jpg"
...
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    PerfumeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        perfumeTitle:
          type: "string"
          example: "Sauvage"
        perfumer:
          type: "string"
          example: "Dior"
        price:
          type: "integer"
          format: "int32"
          example: 62
        filename:
          type: "string"
          example: "Dior_Sauvage.jpg"
        perfumeRating:
          type: "integer"
          format: "int32"
          example: 5
...